test: getProduct에 대한 단위테스트 작성 #2
- JSONError 타입 LocalizedError 에서 Error으로 변경
