+++ OpenMarket/OpenMarketTests/Tests/ParsableTests.swift
@@ -8,7 +8,7 @@
 import XCTest
 @testable import OpenMarket
 
-class ParsableTests: XCTestCase {
+final class ParsableTests: XCTestCase {
     var sut: Parsable!
     
     override func setUpWithError() throws {
@@ -101,20 +101,5 @@ class ParsableTests: XCTestCase {
         }
         
         return page
-    }
-}
-
-extension Product: Equatable {
-    public static func == (lhs: Product, rhs: Product) -> Bool {
-        return lhs.id == rhs.id &&
-            lhs.vendorID == rhs.vendorID &&
-            lhs.name == rhs.name &&
-            lhs.currency == rhs.currency &&
-            lhs.price == rhs.price &&
-            lhs.bargainPrice == rhs.bargainPrice &&
-            lhs.discountedPrice == rhs.discountedPrice &&
-            lhs.stock == rhs.stock &&
-            lhs.createdAt == rhs.createdAt &&
-            lhs.issuedAt == rhs.issuedAt
     }
 }
